refactor: global filters
