+++ src/components/coin/TwitterList.vue
@@ -78,7 +78,9 @@ loadSocial();
 
             <footer class="d-flex flex-wrap align-items-center">
               <p class="mr-3 mb-1 small text-gray">
-                <time :datetime="item.date">{{ date(item.date) }}</time>
+                <time :datetime="item.date">
+                  {{ $filters.date(item.date) }}
+                </time>
               </p>
               <p class="mr-3 mb-1 small text-gray">
                 <b>{{ item.retweet_count }}</b>
