+++ src/types/filters.ts
@@ -1,0 +1,9 @@
+import type { MixinsInterface } from "../helpers/filters";
+
+declare module "@vue/runtime-core" {
+  interface ComponentCustomProperties {
+    $filters: MixinsInterface;
+  }
+}
+
+export {};
